Add mode to simulate the hardware
MockHardwareSetup inherits PurifierHardwareSetup and overwrites the
static functions. All calls to hardware changed to store values with
return parameter mimicing the original call
The calls to hardware are redirected to mock classes for handling
HardwareController updated to call MockHardwareSetup if run_simulator
is set to True.
SimulatorInterface inherits DeviceInterface
SimulatorInterface sends signal to fake the device for ControllerInterface
decive signalling check
SimulatorInterface calls the HardwareController with run_simulator as True